+++ apps/cowswap-frontend/src/modules/limitOrders/containers/LimitOrdersWidget/hooks/useLimitOrdersWidgetActions.ts
@@ -37,7 +37,7 @@ export function useLimitOrdersWidgetActions(): TradeWidgetActions {
 
       const value = tryParseCurrencyAmount(typedValue, currency)
 
-      debouncedTradeAmountAnalytics([field, Number(typedValue)])
+      debouncedTradeAmountAnalytics(field, Number(typedValue))
 
       updateCurrencyAmount({
         activeRate,