+++ apps/explorer/src/hooks/useSearchSubmit.ts
@@ -1,11 +1,14 @@
 import { useCallback } from 'react'
 
+import { useFeatureFlags } from '@cowprotocol/common-hooks'
+import { isTwapEventId } from '@cowprotocol/common-utils'
+
 import { useNavigate } from 'react-router'
 import { useNavigationPathPrefix, useNetworkId } from 'state/network'
 import { Network } from 'types'
 import { isAnAddressAccount, isAnOrderId, isATxHash, isEns } from 'utils'
 
-export function pathAccordingTo(query: string, networkId?: Network | null): string {
+export function pathAccordingTo(query: string, networkId?: Network | null, isTwapEnabled = false): string {
   if (isAnAddressAccount(query)) {
     return 'address'
   }
@@ -16,6 +19,9 @@ export function pathAccordingTo(query: string, networkId?: Network | null): stri
   if (isATxHash(query)) {
     return 'tx'
   }
+  if (isTwapEnabled && isTwapEventId(query)) {
+    return 'twap'
+  }
 
   return 'search'
 }
@@ -24,20 +30,23 @@ export function useSearchSubmit(): (query: string) => void {
   const navigate = useNavigate()
   const prefixNetwork = useNavigationPathPrefix()
   const networkId = useNetworkId()
+  const { isTwapEoaEnabled } = useFeatureFlags()
 
   return useCallback(
     (query: string) => {
       // For now assumes /orders/ path. Needs logic to try all types for a valid response:
       // Orders, transactions, tokens, batches
-      const path = pathAccordingTo(query, networkId)
+      const path = pathAccordingTo(query, networkId, isTwapEoaEnabled)
       const pathPrefix = prefixNetwork ? `${prefixNetwork}/${path}` : `${path}`
 
       if (path === 'address' && isEns(query)) {
         navigate(`/${path}/${query}`)
+      } else if (path === 'twap') {
+        navigate(`/${pathPrefix}/${query}`, { state: { twapGlobalSearch: true } })
       } else {
         query && query.length > 0 && navigate(`/${pathPrefix}/${query}`)
       }
     },
-    [navigate, prefixNetwork, networkId],
+    [isTwapEoaEnabled, navigate, prefixNetwork, networkId],
   )
 }