+++ libs/core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [3.9.1](https://github.com/cowprotocol/cowswap/compare/core-v3.9.0...core-v3.9.1) (2026-09-11)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cowprotocol/common-utils@3.15.1
+
 ## [3.9.0](https://github.com/cowprotocol/cowswap/compare/core-v3.8.0...core-v3.9.0) (2026-09-09)
 
 ### Minor Changes