+++ libs/snackbars/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [2.3.3](https://github.com/cowprotocol/cowswap/compare/snackbars-v2.3.2...snackbars-v2.3.3) (2026-09-11)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cowprotocol/common-hooks@3.11.1
+  - @cowprotocol/ui@3.13.1
+
 ## [2.3.2](https://github.com/cowprotocol/cowswap/compare/snackbars-v2.3.1...snackbars-v2.3.2) (2026-09-09)
 
 ### Patch Changes